+++ utils/align_fingerprints.h
@@ -12,7 +12,7 @@ extern "C" {
 size_t AlignFingerprints(
     const uint32_t *terms1, size_t num_terms1,
     const uint32_t *terms2, size_t num_terms2,
-    int *offsets, int *scores, size_t max_offsets,
+    int *offsets, float *scores, size_t max_offsets,
     int max_distance
 );
